Dashboards: Move empty state button hooks to dedicated file
